pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "anthropic>=0.49.0",
    "autogen-agentchat>=0.4.9.2",
    "autogen-ext[grpc,mcp,ollama,openai]>=0.4.9.2",
    "bs4>=0.0.2",
    "crewai>=0.118.0",
    "gradio>=5.22.0",
    "httpx>=0.28.1",
    "ipywidgets>=8.1.5",
    "langchain>=0.3.25",
    "langchain-anthropic>=0.3.10",
    "langchain-chroma>=0.2.4",
    "langchain-community>=0.3.20",
    "langchain-core>=0.3.63",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.1.5",
    "langchain-groq>=0.3.2",
    "langchain-openai>=0.3.9",
    "langgraph>=0.3.18",
    "langgraph-checkpoint-sqlite>=2.0.6",
    "langsmith>=0.3.18",
    "lxml>=5.3.1",
    "matplotlib>=3.10.3",
    "mcp-server-fetch>=2025.1.17",
    "mcp[cli]>=1.5.0",
    "nbformat>=5.10.4",
    "numpy>=2.2.6",
    "openai>=1.68.2",
    "openai-agents>=0.0.15",
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
    "pip>=25.1.1",
    "playwright>=1.51.0",
    "plotly>=6.0.1",
    "polygon-api-client>=1.14.5",
    "psutil>=7.0.0",
    "pymupdf>=1.26.1",
    "pypdf>=5.4.0",
    "pypdf2>=3.0.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "scikit-learn>=1.7.0",
    "semantic-kernel>=1.25.0",
    "sendgrid>=6.11.0",
    "setuptools>=78.1.0",
    "smithery>=0.1.0",
    "speedtest-cli>=2.1.3",
    "tiktoken>=0.9.0",
    "wikipedia>=1.4.0",
]

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]
